track and abort sessions
